# Conflicts:
#	Modules/Core/Entities/Sale.php
#	Modules/Core/Services/SaleService.php
#	Modules/Finances/Http/Controllers/FinancesApiController.php
#	Modules/Profile/Http/Controllers/ProfileApiController.php
#	Modules/Profile/Transformers/UserResource.php
#	Modules/Withdrawals/Http/Controllers/WithdrawalsApiController.php
#	composer.json


Former-commit-id: bde6832fdd8379194b279921e67241df9c92546a

+++ Modules/Profile/Http/Controllers/ProfileApiController.php
@@ -127,6 +127,10 @@ class ProfileApiController
                 if ($userPhoto != null) {
 
                     try {
+                        if (empty($requestData['photo_w']) || empty($requestData['photo_h'])
+                            || empty($requestData['photo_x1']) || empty($requestData['photo_y1'])) {
+                            return response()->json(['message' => 'Erro ao salvar foto'], 400);
+                        }
                         $digitalOceanService = app(DigitalOceanFileService::class);
                         $digitalOceanService->deleteFile($user->photo);
